requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langgraph>=0.0.20
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
twilio>=8.0.0
email-validator>=2.0.0
fpdf2>=2.7.0
jinja2>=3.1.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
requests>=2.30.0
